requirements.txt
setuptools~=50.3.2
attrs
cattrs